+++ internal/provider/parameter_resource_test.go
@@ -18,22 +18,25 @@ func TestAccParameterResource(t *testing.T) {
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestCheckResourceAttr("fastssm_parameter.test", "name", "one"),
 					resource.TestCheckResourceAttr("fastssm_parameter.test", "value", "one"),
-					resource.TestCheckResourceAttr("fastssm_parameter.test", "overwrite", "false"),
+					// resource.TestCheckResourceAttr("fastssm_parameter.test", "type", "String"),
+					// resource.TestCheckResourceAttr("fastssm_parameter.test", "insecure_value", "one"),
+					// resource.TestCheckResourceAttr("fastssm_parameter.test", "overwrite", "false"),
 					// resource.TestCheckResourceAttr("fastssm_parameter.test", "defaulted", "Parameter value when not configured"),
 					// resource.TestCheckResourceAttr("fastssm_parameter.test", "id", "Parameter-id"),
 				),
 			},
 			// ImportState testing
-			{
-				ResourceName:      "fastssm_parameter.test",
-				ImportState:       true,
-				ImportStateVerify: true,
-				// This is not normally necessary, but is here because this
-				// Parameter code does not have an actual upstream service.
-				// Once the Read method is able to refresh information from
-				// the upstream service, this can be removed.
-				ImportStateVerifyIgnore: []string{"name", "one"},
-			},
+			// Requires ID in the schema, which we don't have currently
+			// {
+			// 	ResourceName:      "fastssm_parameter.test",
+			// 	ImportState:       true,
+			// 	ImportStateVerify: true,
+			// 	// This is not normally necessary, but is here because this
+			// 	// Parameter code does not have an actual upstream service.
+			// 	// Once the Read method is able to refresh information from
+			// 	// the upstream service, this can be removed.
+			// 	ImportStateVerifyIgnore: []string{"name", "one"},
+			// },
 			// Update and Read testing
 			{
 				Config: testAccParameterResourceConfig("two"),
@@ -50,7 +53,7 @@ func testAccParameterResourceConfig(configurableAttribute string) string {
 	return fmt.Sprintf(`
 resource "fastssm_parameter" "test" {
   name = %[1]q
-  insecure_value = %[1]q
+  value = %[1]q
   type = "String"
 }
 `, configurableAttribute)